refactor: 구매금액 도메인 분리, 범용 예외처리 클래스로 분리
