+++ src/main/java/step2/controller/Main.java
@@ -1,6 +1,7 @@
 package step2.controller;
 
 import step2.model.Lotto;
+import step2.model.PurchaseAmount;
 import step2.view.InputView;
 import step2.view.ResultView;
 
@@ -9,8 +10,10 @@ import java.util.Map;
 
 public class Main {
     public static void main(String[] args) {
-        InputView user = new InputView();
-        String input = user.ask("구입금액을 입력해 주세요.");
+        InputView inputView = new InputView();
+        String input = inputView.ask("구입금액을 입력해 주세요.");
+
+        PurchaseAmount user = new PurchaseAmount();
         int amount = user.getPurchaseAmount(input);
         int lottoCount = user.getLottoCount(input);
         System.out.println(lottoCount + "개를 구매했습니다.");
@@ -20,12 +23,12 @@ public class Main {
 
         ResultView view = new ResultView();
         view.showLotto(lottoList);
-
-        input = user.ask("지난 주 당첨 번호를 입력해 주세요.");
-        List<Integer> winningNoList = user.getWinningNumbers(input);
-        Map<Integer, Integer> result = lotto.getWinningResult(lottoList, winningNoList);
-        String winningRate = lotto.getWinningRate(amount, result);
-        view.showResult(result, winningRate);
+//
+//        input = inputView.ask("지난 주 당첨 번호를 입력해 주세요.");
+//        List<Integer> winningNumbers = user.getWinningNumbers(input);
+//        Map<Integer, Integer> result = lotto.getWinningResult(lottoList, winningNumbers);
+//        String winningRate = lotto.getWinningRate(amount, result);
+//        view.showResult(result, winningRate);
 
     }
 }
